feat: split filter
