+++ .gitignore
@@ -5,3 +5,4 @@ tmp/
 *.log
 coverage/
 bun.lockb
+*.bak
